The blacklist and response commands inherit from the same class.
Given that both superclasses are almost identical, it makes sense to
merge them into one.
