+++ zav.discord.blanc.runtime/src/main/java/zav/discord/blanc/runtime/command/mod/ResponseAddCommand.java
@@ -1,6 +1,10 @@
 package zav.discord.blanc.runtime.command.mod;
 
+import jakarta.persistence.EntityManager;
+import java.util.EnumSet;
+import java.util.Set;
 import java.util.regex.Pattern;
+import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.events.interaction.SlashCommandEvent;
 import org.apache.commons.lang3.StringUtils;
 import zav.discord.blanc.api.util.AutoResponseCache;
@@ -12,7 +16,7 @@ import zav.discord.blanc.databind.GuildEntity;
  * This command allows the user to register automatic responses. The bot will respond to any message
  * matching the registered expressions with the pre-defined response.
  */
-public class ResponseAddCommand extends AbstractResponseCommand {
+public class ResponseAddCommand extends AbstractDatabaseCommand {
   private final AutoResponseCache cache;
   
   /**
@@ -27,7 +31,7 @@ public class ResponseAddCommand extends AbstractResponseCommand {
   }
 
   @Override
-  protected String modify(GuildEntity entity, SlashCommandEvent event) {
+  protected String modify(EntityManager entityManager, GuildEntity entity) {
     String pattern = event.getOption("pattern").getAsString();
     String answer = event.getOption("answer").getAsString();
     
@@ -43,6 +47,11 @@ public class ResponseAddCommand extends AbstractResponseCommand {
     cache.invalidate(guild);
 
     return getMessage("response_added", pattern, answer);
+  }
+  
+  @Override
+  protected Set<Permission> getPermissions() {
+    return EnumSet.of(Permission.MESSAGE_MANAGE);
   }
 
 }
